# internal/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /v2/guild-summary:
    get:
      description: Returns clan and member activity totals for a Discord guild.
      parameters:
      - description: Discord guild ID
        in: query
        name: guild_id
        required: true
        type: integer
      - description: Days without login to be considered inactive (default 7)
        in: query
        name: inactive_threshold_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get guild activity summary
      tags:
      - Activity & Inactivity
...
  /v2/inactive-players:
    get:
      description: Returns players in the guild that are considered inactive by the
        provided threshold.
      parameters:
      - description: Discord guild ID
        in: query
        name: guild_id
        required: true
        type: integer
      - description: Days before a player is considered inactive
        in: query
        name: inactive_threshold_days
        type: integer
      - description: Minimum town hall level
        in: query
        name: min_townhall
        type: integer
      - description: Restrict results to a single clan
        in: query
        name: clan_tag
        type: string
      - description: Maximum number of results
        in: query
        name: limit
        type: integer
      - description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get inactive players
      tags:
      - Activity & Inactivity
...
swagger: "2.0"